add chktopo option
it check that at least one topo surface is selected

topo surface defined as the surface that has the max val in

vertical_centerpoint = cubit.get_center_point("surface", k)[2]
vertical_zmax_box=cubit.get_bounding_box('surface',k)[7]
normal=cubit.get_surface_normal(k)

so if only one of the previous condition is not satisfied the surface
is not selected. this could be a problem for a thin layer.